requirements.txt
boto==2.49.0
bpython==0.22.1
django_pdb==0.6.2
feedparser==6.0.10
ipdb==0.13.9
ipython==8.4.0
jupyterlab==3.4.5
mongoengine==0.24.2
MySQL-python==1.2.5
notebook==6.4.12
openpyxl==3.0.10
pip==22.0.2
prompt_toolkit==3.0.30
psycopg2==2.9.3
ptpython==3.0.20
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.11.2
pygraphviz==1.9
pyOpenSSL==22.0.0
PyYAML==6.0
requests==2.25.1
shortuuid==1.0.9
sqlparse==0.4.2
vobject==0.9.6.1
Werkzeug==2.0.2
xhtml2pdf==0.2.8
